+++ iree/task/BUILD
@@ -48,6 +48,7 @@ cc_library(
         "task.c",
         "task_impl.h",
         "topology.c",
+        "topology_cpuinfo.c",
         "worker.c",
         "worker.h",
     ],
@@ -61,6 +62,7 @@ cc_library(
         "submission.h",
         "task.h",
         "topology.h",
+        "topology_cpuinfo.h",
         "tuning.h",
     ],
     deps = [
@@ -164,6 +166,17 @@ cc_test(
 cc_test(
     name = "topology_test",
     srcs = ["topology_test.cc"],
+    deps = [
+        ":task",
+        "//iree/base",
+        "//iree/testing:gtest",
+        "//iree/testing:gtest_main",
+    ],
+)
+
+cc_test(
+    name = "topology_cpuinfo_test",
+    srcs = ["topology_cpuinfo_test.cc"],
     tags = [
         "noasan",  # TODO(benvanik): Does not work on machines with large numbers of cores.
     ],
@@ -172,6 +185,5 @@ cc_test(
         "//iree/base",
         "//iree/testing:gtest",
         "//iree/testing:gtest_main",
-        "@cpuinfo",
     ],
 )
